+++ barfit/data/kinematics.py
@@ -19,9 +19,13 @@ model.
 from IPython import embed
 
 import numpy as np
+from scipy import sparse
 
+from .fitargs import FitArgs
 
-class Kinematics:
+from ..models.beam import construct_beam
+
+class Kinematics(FitArgs):
     r"""
     Base class to hold data fit by the kinematic model.
 
@@ -140,7 +144,7 @@ class Kinematics:
     """
     def __init__(self, vel, vel_ivar=None, vel_mask=None, x=None, y=None, sb=None, sb_ivar=None,
                  sb_mask=None, sig=None, sig_ivar=None, sig_mask=None, sig_corr=None, psf=None,
-                 aperture=None, binid=None, grid_x=None, grid_y=None):
+                 aperture=None, binid=None, grid_x=None, grid_y=None, reff=None):
 
         # Check shape of input arrays
         self.nimg = vel.shape[0]
@@ -160,9 +164,8 @@ class Kinematics:
 
         # Basic properties
         self.spatial_shape = vel.shape
-
-        self.psf = psf
-        self.aperture = aperture
+        self._set_beam(psf, aperture)
+        self.reff = reff
 
         # Build coordinate arrays
         if x is None:
@@ -229,17 +232,39 @@ class Kinematics:
             # bin numbers are not sequential, i.e., the bin numbers are
             # not identical to np.arange(nbin).
 
-            # Construct the bin transform
-            # TODO: Might be a faster way of doing this...
-            # TODO: Set this to be a scipy.sparse.csr_matrix?
-            self.bin_transform = np.array([(binid_map == b).astype(float) 
-                                            for b in self.binid])/nbin.astype(float)[:,None]
+            # Construct the bin transform using a sparse matrix
+            d,i,j = np.array([[1/nbin[i],i,j] 
+                             for i,b in enumerate(self.binid)
+                             for j in np.where(binid_map == b)[0]]).T
+            self.bin_transform = sparse.coo_matrix((d,(i.astype(int),j.astype(int))),
+                                                   shape=(self.binid.size,
+                                                          np.prod(self.spatial_shape))).tocsr()
 
         # Unravel and select the valid values for all arrays
         for attr in ['x', 'y', 'sb', 'sb_ivar', 'sb_mask', 'vel', 'vel_ivar', 'vel_mask', 'sig', 
                      'sig_ivar', 'sig_mask']:
             if getattr(self, attr) is not None:
                 setattr(self, attr, getattr(self, attr).ravel()[indx])
+
+    def _set_beam(self, psf, aperture):
+        """
+        Construct the beam and beam FFT based on the input. If no psf
+        or aperture are provided, both are set to None.
+        """
+        if psf is None and aperture is None:
+            self.beam = None
+            self.beam_fft = None
+            return
+        if psf is None:
+            self.beam = aperture
+            self.beam_fft = np.fft.fftn(np.fft.ifftshift(aperture))
+            return
+        if aperture is None:
+            self.beam = psf
+            self.beam_fft = np.fft.fftn(np.fft.ifftshift(psf))
+            return
+        self.beam_fft = construct_beam(psf, aperture, return_fft=True)
+        self.beam = np.fft.ifftn(self.beam_fft).real
 
     def _ingest(self, data, ivar, mask):
         """
@@ -261,19 +286,29 @@ class Kinematics:
 
         # Set the error and incorporate the mask for a masked array
         if ivar is None:
-            _ivar = np.ones(self.spatial_shape, dtype=float)
+            # Don't instantiate the array if we don't need to.
+            _ivar = None #np.ones(self.spatial_shape, dtype=float)
         elif isinstance(ivar, np.ma.MaskedArray):
             _mask |= np.ma.getmaskarray(ivar)
             _ivar = ivar.data
         else:
             _ivar = ivar
         # Make sure to mask any measurement with ivar <= 0
-        _mask |= np.logical_not(_ivar > 0)
+        if _ivar is not None:
+            _mask |= np.logical_not(_ivar > 0)
 
         return _data, _ivar, _mask
 
+    def remap_data(self, data, masked=False):
+        if data.shape != self.vel.shape:
+            raise ValueError('To remap, must have the same shape as the internal data attributes.')
+        _data = np.ma.masked_all(self.spatial_shape, dtype=float) \
+                    if masked else np.zeros(self.spatial_shape, dtype=float)
+        _data[np.unravel_index(self.grid_indx, self.spatial_shape)] = data[self.bin_inverse]
+        return _data
+
     # TODO: include sigma correction when attr='sig'?
-    def remap(self, attr, masked=True):
+    def remap(self, attr, masked=True, new_attr=True):
         """
         Remap the requested attribute to the full 2D array.
 
@@ -284,6 +319,8 @@ class Kinematics:
                 If an associated mask exists for the selected
                 attribute, return the map as a
                 `numpy.ma.MaskedArray`_.
+            new_attr (:obj:`bool`, optional):
+                make a new attribute `attr_r` for the 2D array.
 
         Returns:
             `numpy.ndarray`_, `numpy.ma.MaskedArray`_: 2D array with
@@ -303,12 +340,15 @@ class Kinematics:
                 = getattr(self, attr)[self.bin_inverse]
         mask_attr = '{0}_mask'.format(attr)
         if not masked or not hasattr(self, mask_attr) or getattr(self, mask_attr) is None:
+            if new_attr: setattr(self, attr+'_r', data)
             return data
 
         mask = np.ones(self.spatial_shape, dtype=bool)
         mask[np.unravel_index(self.grid_indx, self.spatial_shape)] \
                 = getattr(self, mask_attr)[self.bin_inverse]
-        return np.ma.MaskedArray(data, mask=mask)
+        masked_data = np.ma.MaskedArray(data, mask=mask)
+        if new_attr: setattr(self, attr+'_r', masked_data)
+        return masked_data
 
     def bin(self, data):
         """
@@ -328,8 +368,9 @@ class Kinematics:
             ValueError:
                 Raised if the shape of the input array is incorrect.
         """
+        # TODO: Speed this up.
         if data.shape != self.spatial_shape:
             raise ValueError('Data to rebin has incorrect shape; expected {0}, found {1}.'.format(
                               self.spatial_shape, data.shape))
-        return np.dot(self.bin_transform, data.ravel())
+        return self.bin_transform.dot(data.ravel())